requirements.txt
pykrx>=1.0.44
finance-datareader>=0.9.50
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=4.9.3
selenium>=4.15.0
webdriver-manager>=4.0.0
# konlpy>=0.6.0  # Java(JDK) 필요 — 미설치 시 공백 분리 폴백 자동 적용
streamlit>=1.29.0
plotly>=5.18.0
python-dotenv>=1.0.0
tqdm>=4.66.0
jsonschema>=4.19.0
pytest>=7.4.0
